Safari-609.1.11.1/LayoutTests/media/modern-media-controls/tracks-panel/tracks-panel-population-expected.txt

Populating a TracksPanel instance.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tracksPanel.presented is true
PASS tracksPanel.parent is container
PASS scrollContainer.children.length is numberOfSections

Hierarchy for section #0
PASS scrollContainer.children[0].element.localName is "section"
PASS scrollContainer.children[0].children.length is 2
PASS scrollContainer.children[0].children[0].element.localName is "h3"
PASS scrollContainer.children[0].children[0].element.textContent is "Title #0"
PASS scrollContainer.children[0].children[1].element.localName is "ul"
PASS scrollContainer.children[0].children[1].children.length is 3
PASS scrollContainer.children[0].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[0].children[1].children[trackIndex].element.textContent is "Track 0x0"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.className is "selected"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.textContent is "Track 0x1"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[0].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[0].children[1].children[trackIndex].element.textContent is "Track 0x2"

Hierarchy for section #1
PASS scrollContainer.children[1].element.localName is "section"
PASS scrollContainer.children[1].children.length is 2
PASS scrollContainer.children[1].children[0].element.localName is "h3"
PASS scrollContainer.children[1].children[0].element.textContent is "Title #1"
PASS scrollContainer.children[1].children[1].element.localName is "ul"
PASS scrollContainer.children[1].children[1].children.length is 5
PASS scrollContainer.children[1].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[1].children[1].children[trackIndex].element.textContent is "Track 1x0"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[1].children[1].children[trackIndex].element.textContent is "Track 1x1"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[1].children[1].children[trackIndex].element.textContent is "Track 1x2"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.className is "selected"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.textContent is "Track 1x3"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.localName is "li"
PASS scrollContainer.children[1].children[1].children[trackIndex].element.className is ""
PASS scrollContainer.children[1].children[1].children[trackIndex].element.textContent is "Track 1x4"

PASS successfullyParsed is true

TEST COMPLETE
